purchase check:storage

+++ e_shop_user.py
@@ -203,6 +203,13 @@ class Customer(EShopUser):
 		:return: True | False
 		"""
 		try:
+			# check storage first
+			e_shop_cursor.execute(f"SELECT storage_count FROM book_info WHERE ISBN='{isbn}'")
+			storage_count = e_shop_cursor.fetchone()[0]
+			if int(storage_count) < int(count):
+				log.fatal(f"Not enough storage for {isbn}, found {storage_count} require {count}")
+				return False
+			# check end
 			self.cursor.execute(f"INSERT INTO purchase_history(ISBN, time, count) VALUES ({isbn}, {purchase_time}, {count})")
 			self.db.commit()
 			self.cursor.execute(f"DELETE FROM shopping_cart WHERE ISBN={isbn}")
@@ -210,6 +217,8 @@ class Customer(EShopUser):
 			e_shop_cursor.execute(f"INSERT INTO shopping_history(isbn, count, time) VALUES({isbn}, {count}, {purchase_time})")
 			book_e_shop.commit()
 			e_shop_cursor.execute(f"UPDATE book_info SET sold_count=sold_count+{count} WHERE ISBN={isbn}")
+			book_e_shop.commit()
+			e_shop_cursor.execute(f"UPDATE book_info SET storage_count=storage_count-{count} WHERE ISBN={isbn}")
 			book_e_shop.commit()
 			return True
 		except mysql.connector.errors.IntegrityError:
